Allow override for Nominatim base URL
Nominatim can be self-hosted or one might want to use a proxy, for
example to implement rate limiting. To support such use cases, I made
the Nominatim base URL a setting. The default value is the same origin
as was hard-coded in the GeoService requests, so this feature is
backwards compatible.

Notably, being able to use a rate-limiting proxy helps users ensure they
adhere to [Nominatim's usage policy][1].

[1]: https://operations.osmfoundation.org/policies/nominatim/