Add app identifier to User-Agent for Nominatim requests
An openstreetmap.org system administrator recommended we include an app
identifier in the User-Agent header. My inference is that they block
based on the User-Agent. However, I did not want to use only the app
identifier so I found the form that Craft uses and copied that,
appending the app identifier.